Support --update-only flag.
When the Tor Browser is already up-to-date, the current tb-updater will still ask users if they want to download and reinstall it.

With this `--update-only` flag turned on, the tb-updater will exit with a message like "Skip downloading: current Tor Browser 11.5 is already up-to-date".

This flag is especially useful when tb-updater is run regularly to keep the Tor Browser up-to-date because it saves user from spending time and resources downloading an already up-to-date Tor Browser.