Make update-torbrowser work with new multi-lang builds
As of v12.0a4 Tor Browser is shipped as a multi-language archive
and language-specific archives are no longer available. This
patch makes this new multi-lang version install properly but does
not handle starting Tor Browser with the correct locale.

I'm not entirely sure how the locale should be handled. Is it
good enough to just use the system language (which Tor Browser
now does by default) or should some other solution be used? If
the system language is good enough, I'd suggest just merging
this change as is and once v12 stable is released TB_LANG
support can be removed.